refactor(app): extract input misc helpers (phase 4d)
Move switch_to_previous_tab, start_switch_to_tab_prompt,
handle_insert_char_prompt, handle_insert_char_editor,
apply_action_as_events, track_cursor_movement, try_route_composite_key
out of input.rs into a new app/input_helpers.rs sibling. ~344 lines.

Tab navigation, character insertion (both in-prompt and in-editor),
event-batched action application, position-history tracking, and
composite-buffer key routing. None of these are part of the central
key/action dispatch chain.

input.rs now contains only the central dispatch surface
(get_key_context + handle_key + handle_action) plus its helpers
were already extracted earlier — 4138 -> 1448 lines (~65% reduction).

2026 tests still pass.
